+++ src/SleetLib/Utility/RetentionUtility.cs
@@ -1,8 +1,10 @@
 using System;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using NuGet.Packaging.Core;
+using NuGet.Versioning;
 
 namespace Sleet
 {
@@ -22,7 +24,7 @@ namespace Sleet
         /// Find the set of packages to prune for a feed.
         /// Feed package ids that should be checked. Any filtering should happen before this.
         /// </summary>
-        public static HashSet<PackageIdentity> GetPackagesToPrune(HashSet<PackageIdentity> feedPackages, HashSet<PackageIdentity> pinnedPackages, int stableVersionMax, int prereleaseVersionMax)
+        public static HashSet<PackageIdentity> GetPackagesToPrune(HashSet<PackageIdentity> feedPackages, HashSet<PackageIdentity> pinnedPackages, int stableVersionMax, int prereleaseVersionMax, int? groupByUniqueReleaseLabelCount = null)
         {
             var toPrune = new HashSet<PackageIdentity>();
 
@@ -33,7 +35,7 @@ namespace Sleet
 
             var lastId = string.Empty;
             var stable = 0;
-            var pre = 0;
+            var labelLookup = new ConcurrentDictionary<string, int>(StringComparer.OrdinalIgnoreCase);
 
             // Loop through the list of packages
             // When the version count has exceeded the max prune the package
@@ -44,14 +46,19 @@ namespace Sleet
                 if (!StringComparer.OrdinalIgnoreCase.Equals(package.Id, lastId))
                 {
                     stable = 0;
-                    pre = 0;
+                    labelLookup.Clear();
                     lastId = package.Id;
                 }
 
                 var prune = false;
                 if (package.Version.IsPrerelease)
                 {
-                    pre++;
+                    // Split prerelease groups by the given number of release labels
+                    // If the number of labels is 0 then this will always return string.Empty and have
+                    // no impact since all prerelease packages will be in the same group.
+                    var labelKey = GetReleaseLabelKey(package.Version, groupByUniqueReleaseLabelCount);
+                    var pre = labelLookup.AddOrUpdate(labelKey, 1, (k, v) => v + 1);
+
                     prune = pre > prereleaseVersionMax;
                 }
                 else
@@ -68,6 +75,16 @@ namespace Sleet
             }
 
             return toPrune;
+        }
+
+        public static string GetReleaseLabelKey(NuGetVersion version, int? count)
+        {
+            if (count == null || count < 1)
+            {
+                return string.Empty;
+            }
+
+            return string.Join(".", version.ReleaseLabels.Take(count.Value)).ToLowerInvariant();
         }
     }
 }