+++ test/SleetLib.Tests/FeedSettingsUtilityTests.cs
@@ -1,9 +1,6 @@
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
-using System.Text;
 using FluentAssertions;
-using Newtonsoft.Json.Linq;
 using Sleet;
 using Xunit;