+++ test/SleetLib.Tests/RetentionSettingsCommandTests.cs
@@ -43,6 +43,45 @@ namespace SleetLib.Tests
                 success.Should().BeTrue();
                 updatedSettings.RetentionMaxStableVersions.Should().Be(10);
                 updatedSettings.RetentionMaxPrereleaseVersions.Should().Be(5);
+                updatedSettings.RetentionGroupByFirstPrereleaseLabelCount.Should().BeNull();
+            }
+        }
+
+        [Fact]
+        public async Task RetentionSettingsCommand_EnableRetention_WithReleaseLabels()
+        {
+            // Arrange
+            using (var packagesFolder = new TestFolder())
+            using (var target = new TestFolder())
+            using (var cache = new LocalCache())
+            {
+                var log = new TestLogger();
+                var fileSystem = new PhysicalFileSystem(cache, UriUtility.CreateUri(target.Root));
+                var settings = new LocalSettings();
+
+                var context = new SleetContext()
+                {
+                    Token = CancellationToken.None,
+                    LocalSettings = settings,
+                    Log = log,
+                    Source = fileSystem,
+                    SourceSettings = new FeedSettings()
+                    {
+                        CatalogEnabled = true,
+                        SymbolsEnabled = true
+                    }
+                };
+
+                await InitCommand.InitAsync(context);
+
+                // Enable retention
+                var success = await RetentionSettingsCommand.RunAsync(context.LocalSettings, context.Source, 10, 5, 2, false, log);
+                var updatedSettings = await FeedSettingsUtility.GetSettingsOrDefault(context.Source, log, context.Token);
+
+                success.Should().BeTrue();
+                updatedSettings.RetentionMaxStableVersions.Should().Be(10);
+                updatedSettings.RetentionMaxPrereleaseVersions.Should().Be(5);
+                updatedSettings.RetentionGroupByFirstPrereleaseLabelCount.Should().Be(2);
             }
         }
 
@@ -74,7 +113,7 @@ namespace SleetLib.Tests
                 await InitCommand.InitAsync(context);
 
                 // Enable retention
-                var success = await RetentionSettingsCommand.RunAsync(context.LocalSettings, context.Source, 10, 5, false, log);
+                var success = await RetentionSettingsCommand.RunAsync(context.LocalSettings, context.Source, 10, 5, 2, false, log);
 
                 // Disable retention
                 success &= await RetentionSettingsCommand.RunAsync(context.LocalSettings, context.Source, -1, -1, true, log);
@@ -84,6 +123,7 @@ namespace SleetLib.Tests
                 success.Should().BeTrue();
                 updatedSettings.RetentionMaxStableVersions.Should().BeNull();
                 updatedSettings.RetentionMaxPrereleaseVersions.Should().BeNull();
+                updatedSettings.RetentionGroupByFirstPrereleaseLabelCount.Should().BeNull();
             }
         }
     }